feat: 재인증 버튼 클릭 시 타이머 초기화
